feat: add last words configuration options and implement related features

+++ src/main/java/plugily/projects/murdermystery/arena/managers/MapRestorerManager.java
@@ -74,7 +74,7 @@ public class MapRestorerManager extends PluginMapRestorerManager {
   public void clearCorpses() {
     if(!arena.getPlugin().getHookManager().isFeatureEnabled(HookManager.HookFeature.CORPSES)) {
       for(Stand stand : arena.getStands()) {
-        if(!stand.getHologram().isDeleted()) {
+        if(stand.getHologram() != null && !stand.getHologram().isDeleted()) {
           stand.getHologram().delete();
         }
         if(stand.getStand() != null) {
@@ -85,7 +85,7 @@ public class MapRestorerManager extends PluginMapRestorerManager {
       return;
     }
     for(Corpse corpse : arena.getCorpses()) {
-      if(!corpse.getHologram().isDeleted()) {
+      if(corpse.getHologram() != null && !corpse.getHologram().isDeleted()) {
         corpse.getHologram().delete();
       }
       if(corpse.getCorpseData() != null) {
